backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
channels
daphne
graphene-django
django-graphql-jwt
python-dotenv
django-tenants
django-filter
django-environ
dj-database-url
celery
redis